+++ src/components/Container/Container.module.scss
@@ -1,0 +1,30 @@
+.container {
+    min-height: 100vh;
+    width: 100%;
+    padding: 30px;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    &[data-theme="dark"] {
+        background: var(--background-dark);
+    }
+    &[data-theme="light"] {
+        background: var(--background-light);
+    }
+
+    @media screen and (min-width: 1366px){
+        display: flex;
+        justify-content: center;
+    }
+
+    &__second {
+        width: 100%;
+
+        @media screen and (min-width: 1200px){
+            width: 1200px;
+        }
+    }
+
+}
