+++ src/components/Container/index.tsx
@@ -1,53 +1,15 @@
-import { Main, NotUser, Section } from './style'
-import { Header } from '../Header'
-import { useCheckStatusApi } from 'hooks/useCheckStatusApi'
-import { useState, useContext } from 'react'
-import { User } from 'components/User'
-import { GitContext } from 'context/GitProvider'
-import { XCircle } from 'phosphor-react'
+import { ReactNode } from 'react'
+import style from './Container.module.scss'
 
-import imgBg from 'assets/img/gitHub-bg.svg'
+interface ContainerProps {
+  theme: string
+  children: ReactNode
+}
 
-export const Container = () => {
-  const [initial, setInitial] = useState<boolean>(true)
-  const [status, setStatus] = useState<boolean>(false)
-  const { user, repos, starreds, getApi } = useContext(GitContext)
-
-  const getData = async (value: string) => {
-    if (value === '') {
-      setInitial(true)
-      return
-    }
-
-    const { isExist } = await useCheckStatusApi(value)
-
-    setStatus(isExist)
-    setInitial(false)
-    getApi(value)
-  }
-
+export const Container = ({ theme, children }: ContainerProps) => {
   return (
-    <Main>
-      <Header onsubmit={getData} />
-
-      <Section image={imgBg}>
-        {initial ? (
-          <></>
-        ) : (
-          <>
-            {status ? (
-              <User user={user!} repos={repos!} starreds={starreds!} />
-            ) : (
-              <>
-                <NotUser>
-                  <XCircle weight="bold" />
-                  Usário não encontrado
-                </NotUser>
-              </>
-            )}
-          </>
-        )}
-      </Section>
-    </Main>
+    <div className={style.container} data-theme={theme}>
+      <div className={style.container__second}>{children}</div>
+    </div>
   )
 }
